+++ app/app.js
@@ -3,6 +3,6 @@ var React = require('react'),
 
 var
     Router = require('react-router').Router,
-    routes = require('./config/routes');
+    routes = require('./config/react-routes');
 
 ReactDOM.render(routes, document.getElementById('app'));
